+++ CHANGES.md
@@ -17,3 +17,5 @@
 - implémentation de la boucle du jeu et de l'affichage de la grille
 - implementation de la fonction de comptage des voisins
 - implémentation de la fonction step
+- implémentation read_int, input utilisateur pour gérer les dimensions du jeu
+- implémentation de l'option --steps